requirements.txt
click >= 4.1
Pygments >= 1.6
prompt_toolkit==1.0.14
sqlparse>=0.2.2,<0.3.0
configobj >= 5.0.5
cli_helpers >= 0.1.0,<=0.2.3
cx_Oracle
pytest
mock
